refactor: extract reusable GlobalSearch component
Move the inline Ctrl/Cmd+K search dialog out of TopBar into a
standalone GlobalSearch component (views/assets/src/components/common/
GlobalSearch.jsx) with a `variant` prop ("topbar" | "frontend").

- TopBar now renders <GlobalSearch variant="topbar" />, dropping ~140
  lines of duplicated search state, keyboard handling, and dialog markup.
- FrontendLayout gains the search via <GlobalSearch variant="frontend" />
  so frontend users get the same global search as the admin topbar.

No behavior change to search logic (projects, task lists, tasks via
admin-topbar-search + tasks endpoints); just deduplication and reuse.
